requirements.txt
PyQt5
opencv-python
python-dotenv